+++ pkg/deployment/resources/secrets.go
@@ -27,12 +27,15 @@ import (
 	"encoding/hex"
 	"time"
 
+	"github.com/arangodb/kube-arangodb/pkg/deployment/pod"
+	"github.com/pkg/errors"
+
 	"github.com/arangodb/kube-arangodb/pkg/util/constants"
 	jg "github.com/dgrijalva/jwt-go"
 	"k8s.io/apimachinery/pkg/api/equality"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	meta "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	api "github.com/arangodb/kube-arangodb/pkg/apis/deployment/v1"
 	"github.com/arangodb/kube-arangodb/pkg/metrics"
@@ -51,6 +54,7 @@ func (r *Resources) EnsureSecrets() error {
 	ns := r.context.GetNamespace()
 	secrets := k8sutil.NewSecretCache(kubecli.CoreV1().Secrets(ns))
 	spec := r.context.GetSpec()
+	status, _ := r.context.GetStatus()
 	deploymentName := r.context.GetAPIObject().GetName()
 	defer metrics.SetDuration(inspectSecretsDurationGauges.WithLabelValues(deploymentName), start)
 	counterMetric := inspectedSecretsCounters.WithLabelValues(deploymentName)
@@ -71,6 +75,13 @@ func (r *Resources) EnsureSecrets() error {
 		counterMetric.Inc()
 		if err := r.ensureTLSCACertificateSecret(secrets, spec.TLS); err != nil {
 			return maskAny(err)
+		}
+	}
+	if spec.RocksDB.IsEncrypted() {
+		if i := status.CurrentImage; i != nil && i.Enterprise && i.ArangoDBVersion.CompareTo("3.7.0") >= 0 {
+			if err := r.ensureEncryptionKeyfolderSecret(secrets, spec.RocksDB.Encryption.GetKeySecretName(), pod.GetKeyfolderSecretName(deploymentName)); err != nil {
+				return maskAny(err)
+			}
 		}
 	}
 	if spec.Sync.IsEnabled() {
@@ -98,7 +109,7 @@ func (r *Resources) EnsureSecrets() error {
 // of the deployment. If not, it will add such a secret with a random
 // token.
 func (r *Resources) ensureTokenSecret(secrets k8sutil.SecretInterface, secretName string) error {
-	if _, err := secrets.Get(secretName, metav1.GetOptions{}); k8sutil.IsNotFound(err) {
+	if _, err := secrets.Get(secretName, meta.GetOptions{}); k8sutil.IsNotFound(err) {
 		// Secret not found, create it
 		// Create token
 		tokenData := make([]byte, 32)
@@ -117,6 +128,28 @@ func (r *Resources) ensureTokenSecret(secrets k8sutil.SecretInterface, secretNam
 	} else if err != nil {
 		// Failed to get secret for other reasons
 		return maskAny(err)
+	}
+	return nil
+}
+
+func (r *Resources) ensureEncryptionKeyfolderSecret(secrets k8sutil.SecretInterface, keyfileSecretName, secretName string) error {
+	keyfile, err := secrets.Get(keyfileSecretName, meta.GetOptions{})
+	if err != nil {
+		return errors.Wrapf(err, "Unable to find original secret")
+	}
+
+	if len(keyfile.Data) == 0 {
+		return errors.Errorf("Missing key in secret")
+	}
+
+	d, ok := keyfile.Data[constants.SecretEncryptionKey]
+	if !ok {
+		return errors.Errorf("Missing key in secret")
+	}
+
+	owner := r.context.GetAPIObject().AsOwner()
+	if err = k8sutil.AppendKeyfileToKeyfolder(secrets, &owner, secretName, d); err != nil {
+		return errors.Wrapf(err, "Unable to create keyfolder secret")
 	}
 	return nil
 }
@@ -155,7 +188,7 @@ func (r *Resources) ensureExporterTokenSecret(secrets k8sutil.SecretInterface, t
 }
 
 func (r *Resources) ensureExporterTokenSecretCreateRequired(secrets k8sutil.SecretInterface, tokenSecretName, secretSecretName string) (bool, bool, error) {
-	if secret, err := secrets.Get(tokenSecretName, metav1.GetOptions{}); k8sutil.IsNotFound(err) {
+	if secret, err := secrets.Get(tokenSecretName, meta.GetOptions{}); k8sutil.IsNotFound(err) {
 		return true, false, nil
 	} else if err == nil {
 		// Check if claims are fine
@@ -192,7 +225,7 @@ func (r *Resources) ensureExporterTokenSecretCreateRequired(secrets k8sutil.Secr
 // ensureTLSCACertificateSecret checks if a secret with given name exists in the namespace
 // of the deployment. If not, it will add such a secret with a generated CA certificate.
 func (r *Resources) ensureTLSCACertificateSecret(secrets k8sutil.SecretInterface, spec api.TLSSpec) error {
-	if _, err := secrets.Get(spec.GetCASecretName(), metav1.GetOptions{}); k8sutil.IsNotFound(err) {
+	if _, err := secrets.Get(spec.GetCASecretName(), meta.GetOptions{}); k8sutil.IsNotFound(err) {
 		// Secret not found, create it
 		apiObject := r.context.GetAPIObject()
 		owner := apiObject.AsOwner()
@@ -214,7 +247,7 @@ func (r *Resources) ensureTLSCACertificateSecret(secrets k8sutil.SecretInterface
 // ensureClientAuthCACertificateSecret checks if a secret with given name exists in the namespace
 // of the deployment. If not, it will add such a secret with a generated CA certificate.
 func (r *Resources) ensureClientAuthCACertificateSecret(secrets k8sutil.SecretInterface, spec api.SyncAuthenticationSpec) error {
-	if _, err := secrets.Get(spec.GetClientCASecretName(), metav1.GetOptions{}); k8sutil.IsNotFound(err) {
+	if _, err := secrets.Get(spec.GetClientCASecretName(), meta.GetOptions{}); k8sutil.IsNotFound(err) {
 		// Secret not found, create it
 		apiObject := r.context.GetAPIObject()
 		owner := apiObject.AsOwner()